refactor(brain): move the request ledger out of the host
The saves, commits, checkpoints, settlements, and marks a run's record
receives live in BrainRequestLedger; the host keeps the conversation's
standing and the turns, and comes back under its size gate.
